Release version 2.3.24-RC1
Source: 2343ebe28904706ba04143f72ada8bc1b4136a9b

Reviewed-on: https://gerrit.bouffalolab.com/c/release/bouffalo_sdk/+/9158
Tested-by: zrrong <zrrong@bouffalolab.com>

+++ components/ipc/libmetal/port_freertos/sys.c
@@ -27,12 +27,12 @@ void sys_irq_disable(unsigned int vector)
 
 void metal_machine_cache_flush(void *addr, unsigned int len)
 {
-    bflb_l1c_dcache_clean_range((uint32_t)addr, len);
+    bflb_l1c_dcache_clean_range(addr, len);
 }
 
 void metal_machine_cache_invalidate(void *addr, unsigned int len)
 {
-    bflb_l1c_dcache_invalidate_range((uint32_t)addr, len);
+    bflb_l1c_dcache_invalidate_range(addr, len);
 }
 
 void metal_generic_default_poll(void)
